Fix URL normalisation
