+++ src/sources/cfg.rs
@@ -1,7 +1,6 @@
-use super::OrgType;
+use super::{normalize_url, OrgType};
 use crate::{cfg, diag::FileId, LintLevel, Spanned};
 use serde::Deserialize;
-use url::Url;
 
 #[derive(Deserialize, Default)]
 pub struct Orgs {
@@ -164,18 +163,6 @@ impl cfg::UnvalidatedConfig for Config {
             allowed_orgs,
             required_git_spec: self.required_git_spec,
         })
-    }
-}
-
-fn normalize_url(url: &mut Url) {
-    let git_extension = ".git";
-    let needs_chopping = url.path().ends_with(&git_extension);
-    if needs_chopping {
-        let last = {
-            let last = url.path_segments().unwrap().last().unwrap();
-            last[..last.len() - git_extension.len()].to_owned()
-        };
-        url.path_segments_mut().unwrap().pop().push(&last);
     }
 }
